Bad slots and same room
Added optimization w.r.t scheduling in bad slots, but it onlys prevents from starting in a bad slot and not ending in a bad slot.
IMPORTANT
The room constraints did not ensure that one event was scheduled in only 1 room for the entirity of its duration. Constraints have been added in the cut and solv procedure to remedy this.